requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
bitstring==3.1.6
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-celery-results==1.2.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-filter==2.2.0
django-guardian==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-guardian==0.3.0
drf-yasg==1.17.1
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.7
MarkupSafe==1.1.1
packaging==20.1
psycopg2-binary==2.8.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
zipp==3.0.0
